+++ app/frontend/state.ts
@@ -278,8 +278,10 @@ const initialState: State = {
   shouldShowSendTransactionModal: false,
   shouldShowDelegationModal: false,
 }
+
 export type SetStateFn = (newState: Partial<State>) => void
 export type GetStateFn = () => State
+export type Store = {getState: GetStateFn; setState: SetStateFn}
 
 export const getSourceAccountInfo = (state: State) => state.accountsInfo[state.sourceAccountIndex]
 export const getActiveAccountInfo = (state: State) => state.accountsInfo[state.activeAccountIndex]
